requirements.txt
PyMuPDF
google-generativeai>=0.7.2
python-dotenv
